updated filesystem + SFTP untested implementation
