+++ mft-fs/main.go
@@ -1,0 +1,182 @@
+package main
+
+import (
+	"bufio"
+	"flag"
+	"fmt"
+	"os"
+	"strings"
+)
+
+// Define the command-line flags
+var (
+	helpFlag    bool
+	initFlag    string
+	destroyFlag string
+	infoFlag    bool
+)
+
+func init() {
+	// Flag definitions with default values and help text
+	flag.BoolVar(&helpFlag, "help", false, "Display help information for the program.")
+	flag.StringVar(&initFlag, "init", "", "Initialize and mount a filesystem at the specified directory.")
+	flag.StringVar(&destroyFlag, "destroy", "", "Unmount and destroy the filesystem at the specified directory.")
+	flag.BoolVar(&infoFlag, "info", false, "Display information about the mounted filesystem.")
+}
+
+func main() {
+	flag.Parse()
+
+	// Handle the --help flag
+	if helpFlag {
+		printHelp()
+		return
+	}
+
+	// Handle the --init flag
+	if initFlag != "" {
+		handleInit(initFlag)
+		return
+	}
+
+	// Handle the --destroy flag
+	if destroyFlag != "" {
+		handleDestroy(destroyFlag)
+		return
+	}
+
+	// Handle the --info flag
+	if infoFlag {
+		handleInfo()
+		return
+	}
+
+	// If no flags are provided, print help by default
+	printHelp()
+}
+
+// printHelp displays the usage information for the CLI tool.
+func printHelp() {
+	fmt.Println("Usage: mftfs [flags]")
+	fmt.Println("A command-line tool to manage your FUSE filesystems.")
+	fmt.Println("\nFlags:")
+	flag.PrintDefaults()
+}
+
+// handleInit guides the user through the filesystem initialization process.
+func handleInit(mountDir string) {
+	reader := bufio.NewReader(os.Stdin)
+
+	fmt.Printf("Initializing filesystem at: %s\n", mountDir)
+
+	// Ask for the type of filesystem
+	fsType := readInput(reader, "Type of filesystem (s3fs, osfs, grpcfs, sftpfs): ")
+
+	switch strings.ToLower(fsType) {
+	case "sftpfs":
+		handleSFTPFSInit(reader)
+	case "s3fs":
+		handleS3FSInit(reader)
+	case "osfs", "grpcfs":
+		fmt.Printf("Support for %s is not yet implemented.\n", fsType)
+	default:
+		fmt.Printf("Unknown filesystem type: %s\n", fsType)
+		os.Exit(1)
+	}
+
+	// NOTE: You would add the logic here to actually mount the filesystem.
+	// For this prototype, we're just handling the user input.
+	fmt.Printf("Filesystem of type '%s' is ready to be mounted at '%s'.\n", fsType, mountDir)
+	fmt.Println("Add your mounting logic here.")
+}
+
+// handleSFTPFSInit prompts for and reads SFTP connection details.
+func handleSFTPFSInit(reader *bufio.Reader) {
+	fmt.Println("\n--- SFTP-FS Configuration ---")
+	host := readInput(reader, "Host: ")
+	port := readInput(reader, "Port: ")
+	username := readInput(reader, "Username: ")
+	password := readInput(reader, "Password: ")
+
+	// Here you would use these variables to create and start the sftpfsManager
+	// For example:
+	// config := sftpfs.SFTPClientConfig{
+	// 	Host: host,
+	// 	Port: convertPortToInt(port),
+	// 	Username: username,
+	// 	Password: password,
+	// }
+	// manager := sftpfs.NewSFTPFSManager(config)
+	// manager.Start()
+}
+
+// handleS3FSInit prompts for and reads S3 connection details.
+func handleS3FSInit(reader *bufio.Reader) {
+	fmt.Println("\n--- S3-FS Configuration ---")
+	bucketName := readInput(reader, "Bucket Name: ")
+	region := readInput(reader, "Region: ")
+	awsKey := readInput(reader, "AWS Key: ")
+	awsSecretKey := readInput(reader, "AWS Secret Key: ")
+	// You would use these variables to configure S3FS
+}
+
+// handleDestroy is a placeholder for the destroy logic.
+func handleDestroy(mountDir string) {
+	fmt.Printf("Attempting to unmount and destroy filesystem at: %s\n", mountDir)
+	fmt.Println("Add your unmounting logic here.")
+}
+
+// handleInfo is a placeholder for the info logic.
+func handleInfo() {
+	fmt.Println("Displaying information about the filesystem.")
+	fmt.Println("Add your info/metrics retrieval logic here.")
+}
+
+// readInput is a helper function to read a line of input from the user.
+func readInput(reader *bufio.Reader, prompt string) string {
+	fmt.Print(prompt)
+	input, _ := reader.ReadString('\n')
+	return strings.TrimSpace(input)
+}
+
+/*package main
+
+import (
+	"context"
+	"flag"
+	"github.com/jacobsa/fuse"
+	"github.com/jacobsa/fuse/fuseutil"
+	"log"
+	"mft-fs/archive/abstractfs"
+	"mft-fs/archive/osfsmanager"
+	"os"
+)
+
+func main() {
+
+	// reading commandline args
+	mountDirectory := flag.String("mountDirectory", "", "mount directory")
+	rootDirectory := flag.String("rootDirectory", "", "root directory")
+	flag.Parse()
+
+	manager := osfsmanager.NewOSFSManager(*rootDirectory)
+	fs, _ := abstractfs.NewAbstractFS(manager)
+	server := fuseutil.NewFileSystemServer(&fs)
+
+	// mount the filesystem
+	cfg := &fuse.MountConfig{
+		ReadOnly:    false,
+		DebugLogger: log.New(os.Stderr, "fuse: ", 0),
+		ErrorLogger: log.New(os.Stderr, "fuse: ", 0),
+	}
+	mfs, err := fuse.Mount(*mountDirectory, server, cfg)
+	if err != nil {
+		log.Fatalf("Mount: %v", err)
+	}
+
+	// wait for it to be unmounted
+	if err = mfs.Join(context.Background()); err != nil {
+		log.Fatalf("Join: %v", err)
+	}
+}
+*/
