+++ mft-fs/sftpfs/sftp_test_server/host_key
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEowIBAAKCAQEApBlB0edbPaslbnaPfNdwy3qWPOBnemLkfw+KWGO3hKPg/Q8z
+J6OmusjSCapjbfMAxhur/XB8pmwbUwE1P3d+6f5hE1DW5qAxc5eU6QPBK4Y/+f6Z
+021W1sKNc2VDy35tTkpJs8qsjgt1EjvOOi6gOM+Pf1e10F+UWpMYjtJo2k4T7gRT
+vBwGikWXo5Pk99Rnyv3H6PxxoXHervkCgHqWAhpxG+yPS/JsShMzURuRcCYUqLi7
+vUMtojy/IRoQEaXwn9WZPM3fCC4NfIpiEM5RQf4Sy6ibfCtXXcdo1J5szF4TuRYA
+of8PoaVmyuwNYUEQAUuNBM7bycBnjB6Bv4xQKwIDAQABAoIBABeLpjWTkDRvh63r
+ZhhAI2eW3dK7Ui2t3eJbXPrkx/LIvuIgmbLw2N4JLZYEdELU8MH8O3fgeYo5jcwh
+qr3CDBxQCybF3Sf3SVEwKUYr6z4vA6P/w5uiHykufjRWDd23dMCciSm6hXrWKwSk
+Pyl7AjNUmN7pO2W8KSG75HStvSzBdy21QV76RcHPt4d3W0H9OsujRyjENqplJsd6
+Wt5lHytz+PrweaQwo4dRqd6P6Cw5pUoJdL9Dv9aoKv9H95FrCh6CczmCNuwopjuu
+fFK0yyMOBbVGOpQJhleHtpM59bdGox6KXdSHyZ5nf4bHNCUycEOy3hwdSJGsZa0F
+UDPBzXECgYEA4CoD7sO5+N9rA55t6qbB0sWtkTlEG4ZMceBx1NTMJjPr2AFM/yMt
+2zjb3RdLFcPiISB503jswqxCKG6lo48htMYqkzWatvBl4aEmUSdpHN4U7tdcV9Vs
+Hmb3Rt4KoEU6/kWw5afkPTxyW5/8I26bk/kX6n3U6AsejIeASB7kqrsCgYEAu2dt
+rsEZ2WUsyP4rJO0iS9G8U0UyRinaIe//Ba9+NTQzIfYCI4EPmnlpEdxHdXLa68uv
+sC4AHX9uo9WBrCKeDf2BpoOg23KLYh5sfeI+8hqsj2Km15slkvcAiK4+EoqVufy5
+8Edx1sloS6PYi1H7dWmzVXT1AYjkp0t7cMSFsVECgYAsoFCuHOsOH7RXQRgSyeX2
+RwJZ0fO8Iw1e8lbXDxIhXN7L3t5sQJ2lcflbV4LML3X6MFPwhLdlvhNhpz10hU8p
+hPX+3oIRf2RY+dL9/wGPDKtQLTGmDsf6eXcq9ZNZ+s5WxpsD7Vuj0T/W+8+fYiq4
+17MqLKjRFT2crnEcaoINdwKBgHGxvglXfac2ioUmLmaotzXoeKjMQkPvADe2srFw
+VtFQafUgmwuPMTg2kFgVnATVOZv0ZxIr1kYPGcCrt5AHMHysYAv1r5FzXLyaWJfG
+uwmGxUn4P+Bub81JrTk+/Mmi6lN70BLGRXpZlrzNGoLrqfMBvZsK6dmUg6Q7VxiD
+IdkBAoGBAKAl81lQTJuzrGoM5a85pehsc8/eUsQ/hqmSefn5hmwjQeTd93AVcLrJ
+g8dcDfWDHSmMhZ4f67LMl3v0aVR9QUbsdexrtlGomZSCLo8fOh+NFim1j0DbQ2DR
+0Ta8DRUSiO8CzYs7Y5zKydwQVhWol6FhCojunc4d09lpZ3i8c36w
+-----END RSA PRIVATE KEY-----
